Add custom auth page
This includes a route at /auth. The route has several components that make up the auth form.

The form is dual-use and allows the user to switch between sign in or register.

The forms include user-facing validation and sign in/register errors. On successful auth, the user is redirect to the home page.