Refactor AddBoard modal to add a random color to each column
This is used for an accent color at the top of each column next to the name. It is only decorative.

Updated related attribute on Column component as well as the type for board/column creation to add a "color" attribute.

Added a utility function and a colors constant from which colors are drawn.