Add custom next middleware for request auth validation
This middleware runs before each request is processed and checks for a valid user and session.

If a protected page is being accessed but there is no valid user, page redirects to /auth.

If the base or auth page are accessed but there is a valid user, page redirects to /boards.

This is processed before the request so this avoids any flashes of unintended pages.