Fix boards search
The previous implementation was missing results if multiple results were found on the same board. Essentially, if a match was found, the search function would move on to the next board.

This fix implements a new search method that looks for individual results instead of board-level results. Results are now displayed in the Search component as individual board, column, task, or subtask items instead of boards that contain those items.

Styling was updated so that search results do not overflow the container.

The searchBoards method was moved to the utils folder.