Refactor search result item component
Because the way search is conducted was changed, this component had to be refactored. It now displays the results prefixed by which entity they're found in: board, column, task, or subtask. New icons were added for this updated UI.

The result preview was also updated to show the area in the description where the query was found.

The getSearchContext method was moved to the utils folder.